refactor: move "sort list and remove duplicates" to new function
