+++ index.js
@@ -12,7 +12,6 @@ import config from './lib/config.js';
 import getDependencies from './lib/getDependencies.js';
 import getFormatter from './lib/getFormatter.js';
 import listToTree from './lib/listToTree.js';
-import { getPackageIdWithVersion } from './lib/packageIdentity.js';
 import util from './lib/util.js';
 
 const debug = createDebugMessages('license-report-recurse');
@@ -120,17 +119,8 @@ const debug = createDebugMessages('license-report-recurse');
      * 'config.only' defaults to 'prod', 'opt', 'peer'
      */
     if (config.output !== 'tree') {
-      const sortedList = depsIndex.sort(util.alphaSort);
-      // remove duplicates as they are only needed to identify dependency loops
-      let lastPackage = '';
-      const dedupedSortedList = sortedList.filter((element) => {
-        const currentPackage = getPackageIdWithVersion(element);
-        if (currentPackage !== lastPackage || element.isRootNode) {
-          lastPackage = currentPackage;
-          return true;
-        }
-        return false;
-      });
+      // remove duplicates as they are not needed for flat output formats
+      const dedupedSortedList = util.dedupeDependenciesList(depsIndex);
 
       // keep only fields that are defined in the configuration
       const packagesList = await Promise.all(
