+++ lib/util.js
@@ -1,6 +1,8 @@
 import fs from 'node:fs';
 import semver from 'semver';
 
+import { getPackageIdWithVersion } from './packageIdentity.js';
+
 /**
  * Get an object by asynchronously reading a file
  * @param {string} path - path of the json file
@@ -74,6 +76,27 @@ const alphaSort = (elementA, elementB) => {
   return 0;
 };
 
+/**
+ * Remove duplicate entries from the list of dependencies; the
+ * list is sorted by package name before removing the duplicates.
+ * @param {Array<object>} dependenciesList - list of dependencies
+ * @returns {Array<object>} sorted and deduped list of dependencies
+ */
+const dedupeDependenciesList = (dependenciesList) => {
+  const sortedList = dependenciesList.sort(alphaSort);
+
+  let lastPackage = '';
+  const dedupedSortedList = sortedList.filter((element) => {
+    const currentPackage = getPackageIdWithVersion(element);
+    if (currentPackage !== lastPackage || element.isRootNode) {
+      lastPackage = currentPackage;
+      return true;
+    }
+    return false;
+  });
+  return dedupedSortedList;
+};
+
 const helpText = `Generate a detailed report of the licenses of all projects dependencies with option recursive.
 
 Usage: license-report-recursive [options]
@@ -111,5 +134,6 @@ export default {
   deepCloneObject,
   isNullOrUndefined,
   alphaSort,
+  dedupeDependenciesList,
   helpText,
 };
